Type overlay 47 actor animation frame as unsigned
The animation-frame byte at actor+0x3B was declared signed and read
signed at both of its two sites: the animation-rate table index and the
six-case dispatch. The target reads it once signed and once unsigned --
one lb and one lbu at that displacement, against our two lb.

The field is unsigned. The dispatch reads it plainly, which is where the
lbu comes from, and the rate-table index takes the signed cast, which is
where the lb comes from. That is the opposite arrangement to the one the
reconstruction had, where a `(u8)` cast on a signed field produced a
second lb and left the switch to mask afterwards.

The displacement census now agrees exactly at that field, and positional
masked words fall 1983 -> 1906, the best measured on this function; the
previous best of 1969 was on the candidate with the four-byte-too-large
player record. Seventy-seven words is far more than the two reads
themselves, so the correction propagates: the dispatch is the hub of the
largest basic-block cluster in the function and its selector's width was
changing the comparison and branch forms downstream of it.

gmake verify prints 507341c0a40ca3e9a7cee969b396ee53facfb548, gmake
cleanroom passes, and gmake check-nonmatching-builds is OK across all
242 TUs.

+++ src/overlays/o047/func_overlay_047_F0000B30_1891948.c
@@ -21,7 +21,7 @@ typedef struct Overlay47Actor {
     u8 pad18[0x10];
     f32 frameValue;
     u8 pad2C[0xF];
-    s8 frame;
+    u8 frame;
     u8 pad3C[0xA];
     s16 kind;
     u8 pad48[0x1C];
@@ -360,7 +360,7 @@ void func_overlay_047_F0000B30_1891948(s32 updateRate) {
             }
             actor = player->actor;
             oldFrame = actor->frameValue;
-            func_8005ABA8(actor, ov47Data_3F0[actor->frame], rate);
+            func_8005ABA8(actor, ov47Data_3F0[(s8)actor->frame], rate);
             scroll = player->actor->scroll;
             scroll->b += speed;
             scroll->a += speed;
@@ -399,7 +399,7 @@ void func_overlay_047_F0000B30_1891948(s32 updateRate) {
             if (actor != NULL) {
                 actor->trigger = 0;
                 actor = player->actor;
-                switch ((u8)actor->frame) {
+                switch (actor->frame) {
                     case 1:
                         frame = actor->frameValue;
                         if (0.3f <= frame && frame < 0.65f) {
